+++ src/functions/arithmetic/modulo.js
@@ -16,8 +16,10 @@ class LimeFunctionModulo extends LimeFunction {
       'tr(_+)',
       'tr(_-)',
 
-      'cl(var->exp)',
-      'cr(var->exp)',
+      'cl(var->expr)',
+      'cr(var->expr)',
+      'cl(bool->int)',
+      'cr(bool->int)',
 
       'eb(int,int)',
     ];