+++ src/functions/arithmetic/positive.js
@@ -16,7 +16,8 @@ class LimeFunctionPositive extends LimeFunction {
       'tr(_+)',
       'tr(_-)',
 
-      'cr(var->exp)',
+      'cr(var->expr)',
+      'cr(bool->int)',
 
       'er(int)',
       'er(rat)',