+++ src/functions/logical/logicalNot.js
@@ -16,7 +16,8 @@ class LimeFunctionLogicalNot extends LimeFunction {
       'tr(_+)',
       'tr(_-)',
 
-      'cr(var->exp)',
+      'cr(var->expr)',
+      'cr(!bool->bool)',
 
       'er(bool)',
     ];